+++ strider/commands.py
@@ -20,50 +20,27 @@ class KeyCommand(Registry):
         updates the __doc__ of a command with the names of the keys bound to it.
         Codes that have no name in the code_dict are ignored.
         """
+        # currently all key command docs are single-line, but let's future-proof
         self.__doc__ = dedent(self.__doc__).strip()
         if self.code_dict:
-            keys = []
-            codes = set(self.key_codes)
             # this seems pretty inefficient, but it's O(n^2) where n = number of key commands
             # doubling the number of commands currently: n=50, O(2500), doesn't even make a dent next to the stuff
             # opencv does
             for k, v in self.code_dict.items():
-                if v in codes:
-                    keys.append(pretty_key_name(k))
-            # currently all key command docs are single-line, but let's future-proof
-            if keys:
-                self.__doc__ = ' or '.join(keys) + ': ' + self.__doc__
+                if v == self.key_code:
+                    self.__doc__ = pretty_key_name(k)+': '+self.__doc__
 
     def register_key(self):
-        # NOTE: key command register_keys are not registered directly!!!
-        return self.key_codes
+        return self.key_code
 
-    def register(self, key=...):
-        if key is ...:
-            key = self.register_key()
-        for k in key:
-            super().register(k)
-
-    @classmethod
-    def values_distinct(cls):
-        # filters out repeating values
-        prev = None
-        for v in super().values():
-            if prev is v:
-                continue
-            prev = v
-            yield v
-
-    def __init__(self, key_codes: Union[Iterable[int], int], func: Callable):
-        if isinstance(key_codes, int):
-            key_codes = (key_codes,)
-
-        self.key_codes = key_codes
+    def __init__(self, key_code: int, func: Callable, *, allow_on_auto_play=False):
+        self.key_code = key_code
         update_wrapper(self, func)
-        self._update_doc()
         self.__func__ = func
+        self._update_doc()
+        self.allow_on_auto_play = allow_on_auto_play
+
         super().__init__()
-        self.allow_on_auto_play = False
 
     def __call__(self, *args, **kwargs):
         return self.__func__(*args, **kwargs)
@@ -71,11 +48,7 @@ class KeyCommand(Registry):
 
 def key_command(codes: Union[Iterable[int], int, str], **kwargs):
     """A KeyCommand wrapper to be used as a decorator"""
-    def ret(func):
-        kc = KeyCommand(codes, func)
-        kc.__dict__.update(kwargs)
-        return kc
-    return ret
+    return partial(KeyCommand, codes, **kwargs)
 
 
 class SpecialCommand(Registry):
@@ -114,6 +87,7 @@ class SpecialCommand(Registry):
         update_wrapper(self, func)
         self.__func__ = func
         self.__doc__ = self.__name__ + str(inspect.signature(self.__func__)) + ': ' + self.__doc__
+
         super().__init__()
 
     def __call__(self, *args, **kwargs):